requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
Django==3.2.8
django-rest-knox==4.1.0
djangorestframework==3.12.4
pycparser==2.20
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
